+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [4.8.2] - 2025-11-24
+### Fixed
+- the ~(equivalence) operator for quantities (now the comparison is performed
+  after converting the values to the most granular unit).
+
 ## [4.8.1] - 2025-11-23
 ### Fixed
 - an issue where the resolve() function does not resolve URL in FHIR.string.